stm32f4-7: partially merged the rcc and moved static inline functions to c
