+++ lib/stm32/f7/Makefile
@@ -55,8 +55,8 @@ OBJS		+= gpio.o gpio_common_all.o gpio_common_f0234.o
 OBJS		+= i2c_common_v2.o
 OBJS		+= iwdg_common_all.o
 OBJS		+= ltdc_common_f47.o 
-OBJS		+= pwr.o rcc.o
-OBJS		+= rcc_common_all.o
+OBJS		+= pwr.o
+OBJS		+= rcc.o rcc_common_f47.o rcc_common_all.o
 OBJS		+= rng_common_v1.o
 OBJS		+= spi_common_all.o spi_common_v2.o
 OBJS		+= timer_common_all.o
